Added ability to calculate total of user basket
Added new method to Shop.cs that will calculate the total of all of the items within the users basket. Does not take into account the promotions yet but will be added next
Also added the ability for a user to choose an item from the shop list and add it to their basket. The user can add as many items as they want at once.
Also added validation for theses inputs.